api/requirements.txt
flask
flask-restx
mysql-connector-python